Move NRTM4 code out of whois-api module and into whois-nrtmv4 module where it belongs
