+++ whois-nrtm4/src/main/java/net/ripe/db/nrtm4/servlet/NrtmExceptionMapper.java
@@ -1,4 +1,4 @@
-package net.ripe.db.whois.api.nrtm4;
+package net.ripe.db.nrtm4.servlet;
 
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.ws.rs.BadRequestException;
